requirements.txt
discord.py==1.3.3
async-timeout==3.0.1
python-dotenv==0.13.0
requests==2.24.0
asyncpg==0.20.1
